Add GitLab Pages and Cloudflare Pages to inline 404 help
I had to move an Eleventy site to Cloudflare Pages and was pleasantly surprised when 404.html worked the same as GitHub Pages and Netlify et al.

I updated the website quicktip (11ty/11ty-website#1608).

Updating the base blog, I noticed the inline link to GitLab Pages was missing, so I added that as well.